+++ examples/using-asciidoc/gatsby-node.js
@@ -2,8 +2,7 @@ const _ = require(`lodash`)
 const Promise = require(`bluebird`)
 const path = require(`path`)
 const slash = require(`slash`)
-const asciidoc = require(`asciidoctor.js`)()
-const crypto = require(`crypto`)
+const { createFilePath } = require(`gatsby-source-filesystem`)
 
 // Implement the Gatsby API “createPages”. This is
 // called after the Gatsby bootstrap is finished so you have
@@ -23,7 +22,9 @@ exports.createPages = ({ graphql, actions }) => {
             edges {
               node {
                 id
-                slug
+                fields {
+                  slug
+                }
               }
             }
           }
@@ -46,7 +47,7 @@ exports.createPages = ({ graphql, actions }) => {
           // as a template component. The `context` is
           // optional but is often necessary so the template
           // can query data specific to each page.
-          path: edge.node.slug,
+          path: edge.node.fields.slug,
           component: slash(articleTemplate),
           context: {
             id: edge.node.id,
@@ -59,36 +60,15 @@ exports.createPages = ({ graphql, actions }) => {
   })
 }
 
-exports.onCreateNode = async ({
-  node,
-  createNodeId,
-  actions,
-  loadNodeContent,
-}) => {
-  const { createNode, createParentChildLink } = actions
-  if (node.extension === `adoc`) {
-    const content = await loadNodeContent(node)
-    const html = asciidoc.convert(content)
-    const slug = `/${path.parse(node.relativePath).name}/`
-    const asciiNode = {
-      id: createNodeId(`${node.id} >>> ASCIIDOC`),
-      parent: node.id,
-      internal: {
-        type: `Asciidoc`,
-        mediaType: `text/html`,
-        content: html,
-      },
-      children: [],
-      html,
-      slug,
-    }
+exports.onCreateNode = async ({ node, actions, getNode, loadNodeContent }) => {
+  const { createNodeField } = actions
 
-    asciiNode.internal.contentDigest = crypto
-      .createHash(`md5`)
-      .update(JSON.stringify(asciiNode))
-      .digest(`hex`)
-
-    createNode(asciiNode)
-    createParentChildLink({ parent: node, child: asciiNode })
+  if (node.internal.type === `Asciidoc`) {
+    const value = createFilePath({ node, getNode })
+    createNodeField({
+      name: `slug`,
+      node,
+      value,
+    })
   }
 }