+++ examples/using-javascript-transforms/src/components/HelmetBlock/index.js
@@ -1,5 +1,5 @@
 import React from "react"
-import Helmet from "react-helmet"
+import { Helmet } from "react-helmet"
 import moment from "moment"
 
 class HelmetBlock extends React.Component {