+++ examples/using-remark/gatsby-config.js
@@ -52,6 +52,13 @@ module.exports = {
               dashes: `oldschool`,
             },
           },
+          {
+            resolve: `gatsby-remark-embed-snippet`,
+            options: {
+              // Example code links are relative to this dir.
+              directory: `${__dirname}/src/code-examples/`,
+            },
+          },
           `gatsby-remark-prismjs`,
           `gatsby-remark-autolink-headers`,
           `gatsby-remark-katex`,